requirements.txt
async-generator==1.10
attrs==23.1.0
autopep8==2.0.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
boto3==1.26.127
botocore==1.29.127
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
outcome==1.2.0
packaging==23.1
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.29.0
s3transfer==0.6.1
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.44
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
vboxapi==1.0
webdriver-manager==3.8.6
Werkzeug==2.2.2
wsproto==1.2.0
